+++ src/application/Base.java
@@ -1,3 +1,10 @@
+package application;
+
+import entities.Aluno;
+import entities.Disciplina;
+import entities.Laboratorio;
+import entities.Professor;
+
 import java.util.ArrayList;
 import java.util.List;                                                        // Powered by: BRUNO CASÉ and ZAIRA DUTRA
                                                                                //Last att: 12:58 pm  (21/11/2023)
