+++ src/auxiliar/Solicitacao.java
@@ -1,3 +1,10 @@
+package auxiliar;
+
+import entities.Aluno;
+import entities.Disciplina;
+import entities.Laboratorio;
+import entities.Professor;
+
 import java.util.List;
                                                                                  // Powered by: BRUNO CASÉ and ZAIRA DUTRA
                                                                                 //Last att: 12:58 pm  (21/11/2023)
@@ -78,7 +85,7 @@ public class Solicitacao {                 //ATRIBUTOS
 
 
     public String toString() {                                   // METODO TO STRING
-        return "Solicitacao{" +
+        return "auxiliar.Solicitacao{" +
                 "id=" + id +
                 ", laboratorio=" + laboratorio +
                 ", professor=" + professor +
